Build and install project gem steps
MSP-11400

'I build gem from project's "<gemspec path>"' will do `gem build
<gemspec path>` where <gemspec_path> is relative to PWD, which is
assumed to the project's root, and not Aruba's current directory.  'I
install latest local "<name>"' looks in the PWD for the lastest gem with
given name and installs it.  It is recommended that 'I use gemset
"<name>"' is used before the installation so that it doesn't affect the
project under test and only exists during the scenario.